url: expose urlToOptions utility
