+++ lib/url.js
@@ -47,7 +47,8 @@ const {
   domainToUnicode,
   formatSymbol,
   pathToFileURL,
-  fileURLToPath
+  fileURLToPath,
+  urlToOptions
 } = require('internal/url');
 
 // Original url.parse() API
@@ -978,5 +979,6 @@ module.exports = {
 
   // Utilities
   pathToFileURL,
-  fileURLToPath
+  fileURLToPath,
+  urlToOptions
 };
